fix(ai): stop the thunk policy pre-check from pulling in aiProviderService.ts's full provider graph
The CI Quality Gate caught a real regression from the previous commit: aiThunkUtils.ts's
policy pre-check imported resolvePositiveRoutingOpts from aiProviderService.ts, a hub module
whose own top-level imports eagerly load every provider adapter (OpenRouter, Gemini, Ollama,
local facades). Any test file that transitively imports an AI thunk and mocks services/logger
with a partial shape (no createLogger export) started failing at module-eval time, since one
of those adapters (openrouterModels.ts) calls createLogger() at module scope. Confirmed via
the exact 3 affected suites in the CI run (AdvancedImportExport, listenerMiddleware,
keyboard/shortcutActions).

Root-cause fix: extracted resolvePositiveRoutingOpts (and its private on-device-provider set)
out of aiProviderService.ts into a new services/ai/positiveRouting.ts, which only imports the
lightweight aiModeService.ts and routingLogger.ts — not any provider adapter. aiThunkUtils.ts
now imports from this module directly instead of from aiProviderService.ts, which still
imports and uses the same function internally for generateText/streamText.

routingLogger.ts itself also calls createLogger()/sanitizeLogContext() at module scope, so the
same 3 test files' logger mocks needed both added (their existing mocks already covered the
plain  export, just not the factory/sanitizer used by createLogger() call sites) --
mirroring the identical fix already established in tests/unit/dbServiceRetry.test.ts for the
same defect class.
